+++ Angular2Emulator/TypeScript/Ng2Emulation/Templates/Rules/NgProperty.ts
@@ -1,7 +1,0 @@
-﻿import {ParserRule} from "../ParserRule";
-
-export default class NgPropertyRule implements ParserRule {
-    processTemplate(template: string): string {
-        return template.replace(/#textInput/g, "ng-property=\"textInput\"");
-    }
-}